Create code
This code defines a Phone class with attributes like model, color, and phoneNumber. It has two member functions, call() to make a call to a given number and displayInfo() to display the phone's information.

In the main() function, the user is prompted to enter the phone's model, color, and number. Then, an instance of the Phone class is created with the provided information. The phone's information is displayed, and the user is prompted to enter a number to call. Finally, the call() function is called with the given number.

Please note that this is a basic example and does not include all the functionalities and error handling you might find in a real phone program.